requirements.txt
nltk==3.4.1
matplotlib==3.0.3
numpy==1.18.1
cycler==0.10.0
vega_datasets==0.7.0
pandas==1.0.3
altair==3.2.0
Pillow==8.0.1
